refactored database classes
moved common methods to utils file
renamed database file to only contain one type
renamed test file to reflect new database name
moved mock methods into test file since they aren't common
